requirements.txt
# PDF Processing
pdfplumber>=0.9.0
pymupdf>=1.23.0  # Correct package name for PyMuPDF, had issues in Linux

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# OCR Support
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.5.0

# Data Export
pyarrow>=14.0.0  # For Parquet support
reportlab>=3.6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
